Finishes group module

+++ src/main/java/post/services/postManagmentService.java
@@ -11,6 +11,8 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import javax.ws.rs.core.Response;
 
+import group.entity.groupEntity;
+import group.interfaces.repositories.IGroupRepository;
 import post.entity.postEntity;
 import post.interfaces.repositories.IPostRepository;
 import post.interfaces.services.IPostManagmentService;
@@ -27,6 +29,8 @@ public class postManagmentService implements IPostManagmentService {
 	IPostRepository postDatabaseManager;
 	@Inject
 	IuserRepository userDatabaseManager;
+	@Inject
+	IGroupRepository groupDatabaseManager;
 
 
 	@Override
@@ -97,9 +101,84 @@ public class postManagmentService implements IPostManagmentService {
     }
 
 	@Override
-	public void delete(int postId) {
+	public Response delete(int postId, HttpServletRequest request) {
+		HttpSession session = request.getSession(false);
+        if (session == null || session.getAttribute("userId") == null) {
+            return Response.status(Response.Status.UNAUTHORIZED)
+                           .entity("User not authenticated.")
+                           .build();
+        }
+
+        int userId = (Integer) session.getAttribute("userId");
+        
+        postEntity post = postDatabaseManager.findById(postId);
+        if (post == null) {
+            return Response.status(Response.Status.NOT_FOUND)
+                           .entity("Post not found.")
+                           .build();
+        }
+
+        if (post.getAuthor().getId() != userId) {
+            return Response.status(Response.Status.FORBIDDEN)
+                           .entity("Not authorized to update this post.")
+                           .build();
+        }
+        
 		postDatabaseManager.delete(postId);
-		
+		return Response.ok()
+                .entity("Post created successfully!")
+                .build();
+	}
+
+
+	@Override
+	public Response createPostInGrp(postEntity newPost, int grpId, HttpServletRequest servlet) {
+		HttpSession session = servlet.getSession(false);
+        if (session == null || session.getAttribute("userId") == null) {
+            return Response.status(Response.Status.UNAUTHORIZED)
+                           .entity("User not authenticated.")
+                           .build();
+        }
+
+        int userId = (Integer) session.getAttribute("userId");
+        userEntity author = userDatabaseManager.findById(userId);
+        if (author == null) {
+            return Response.status(Response.Status.BAD_REQUEST)
+                           .entity("Invalid user.")
+                           .build();
+        }
+        
+        boolean check = false;
+        for(groupEntity grp : author.getGroups()) {
+        	if(grp.getGrpId() == grpId) {
+        		check = true;
+        		break;
+        	}
+        }
+        
+        if(check) {
+        	groupEntity group = groupDatabaseManager.getGroupById(grpId);
+            if (group == null) {
+                return Response.status(Response.Status.BAD_REQUEST)
+                               .entity("Group not found.")
+                               .build();
+            }
+            
+            newPost.setAuthor(author);
+            newPost.setGroup(group);
+            postDatabaseManager.createPost(newPost);
+            author.getPosts().add(newPost);
+            group.getAllPostsInGrp().add(newPost);
+            return Response.status(Response.Status.CREATED)
+                    .entity("Post created in group successfully!")
+                    .build();
+        }
+        else {
+        	return Response.status(Response.Status.BAD_REQUEST)
+                    .entity("You are not a member in the group!")
+                    .build();
+
+        }
 	}
